Alberto Vignani: hppa.cfront CC and gcc-2.6.4 failed to compile this file since Fraction now has an int method. ../../../../src/domains/sdf/loopScheduler/ LSGraph.cc: In method `void LSGraph::sortedInsert(class LSNode *)': ../../../../src/domains/sdf/loopScheduler/ LSGraph.cc:76: ambiguous default type conversion for `operator >=' make[2]: *** [LSGraph.o] Error 1 The fix is to add a (int) cast.

+++ ptclassic/src/domains/sdf/loopScheduler/LSGraph.cc
@@ -73,7 +73,7 @@ void LSGraph::sortedInsert(LSNode *p) {
 
 	while ((nl = nextNode.nextLink()) != 0) {
 		LSNode* n = nl->node();
-		if (freq >= n->myMaster()->repetitions) {
+		if (freq >= (int)(n->myMaster()->repetitions)) {
 			candidates.insertAhead(tmp,nl);
 			return;
 		}
